26 create account login (#10)
* feat: Added user registration and login endpoints with firebase auth"

Still need to bugtest and properly validate email addresses.

* feat: Added frontend to test signup/login

Added texfields to the urls http://localhost:5173/login and http://localhost:5173/sign-up so that you can sign up and login.

* feat: Sign in and Sign up in moved to frontend

The actual sign up and sign in with firebase is now done in the frontend and a token is sent to the backend to verify the session.

Verification email is working as well, but its buggy.

* feat: Done styling Login

Done styling the login page to look like the figma.

* feat: Done styling for sign up

Finished styling the signup page

* chore: removed email verification

Had issues with checking if they clicked the verification email on the frontend.

* feat: Added back email verification

Fixed email verification and redirection to the /verification-instructions page

* feat: Clicking QuickBytes title takes you back to root page

Clicking QuickBytes in the navbar takes you to the home page /

* chore: Put important backend values into env file

In the frontend the firebase client sdk is initialized but the environment variables aren't working properly so its hardcoded.

* chore: removed alert

Removed an alert that was added to the login page.

* chore: Added Comment Code

Added comment code to describe the files. Removed unnecessary comments.

* chore: merged develop App.tsx into feature branch

* fix: Used wrong method to get database

Used wrong method to access get database in backend.

* chore: Put firebase config values in env file

Put firebase config values into environment variables file.

* chore: Added firebase-admin to backend

* chore: Added firebase dependency to frontend

* fix: Add item to order

* fix: Fixed addItemToOrder

* feat: Protected other pages

* styling: Changed styling for Main.tsx

Changed the background color to white

* feat: Make sure emails are u of t emails

* chore: Removed unnecessary comments

* feat: Navbar signin/signout button is based on login state

This way the user cannot see the sign in button when they are already signed in if they navigate to the root page.

* chore: Single Navbar which changes based on signin state